requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.1
helicone==1.0.13
PyPDF2==3.0.1
Pillow==10.2.0
fpdf2==2.7.8
dateparser==1.1.8 